-Added a New File button

+++ NotepadSharp/DocumentView/DocumentViewModel.cs
@@ -3,40 +3,53 @@ using System.Linq;
 using System.IO;
 using Utility;
 using WPFUtility;
+using Microsoft.Win32;
 
 namespace NotepadSharp {
     //This view model deals with abstracting the persistency away from the content vm
     public class DocumentViewModel : ViewModelBase {
         SerializableFileInfo _fileInfo;
+        Action<string> _updateLabelCallback;
 
-        public DocumentViewModel(string filePath) {
-            _fileInfo = ArgsAndSettings.CachedFiles.FirstOrDefault(x => x.OriginalFilePath == filePath);
+        public DocumentViewModel(string filePath, Action<string> updateLabelCallback) {
+            _updateLabelCallback = updateLabelCallback;
+            _fileInfo = ArgsAndSettings.CachedFiles.FirstOrDefault(x => x.OriginalFilePath == filePath || x.CachedFilePath == filePath);
             IsDirty = new NotifyingProperty<bool>(x => SaveFileInfo());
 
             if(_fileInfo == null) {
                 var uid = Guid.NewGuid().ToString();
+                var fileName = Path.GetFileName(filePath);
+                var cachedDirectoryPath = Path.Combine(Constants.FileCachePath, uid);
+
                 _fileInfo = new SerializableFileInfo();
-                _fileInfo.CachedFilePath = Path.Combine(Constants.FileCachePath, uid);
-                _fileInfo.IsDirty = false;
-                _fileInfo.OriginalFilePath = filePath;
-                _fileInfo.Hash = Methods.HashFile(filePath);
-                File.Copy(filePath, _fileInfo.CachedFilePath);
+                _fileInfo.CachedFilePath = Path.Combine(cachedDirectoryPath, fileName);
+                Directory.CreateDirectory(cachedDirectoryPath);
+
+                if(File.Exists(filePath)) {
+                    _fileInfo.OriginalFilePath = filePath;
+                    _fileInfo.Hash = Methods.HashFile(filePath);
+                    File.Copy(filePath, _fileInfo.CachedFilePath);
+                } else {
+                    File.WriteAllText(_fileInfo.CachedFilePath, "");
+                }
+
                 ArgsAndSettings.CachedFiles.Add(_fileInfo);
-            } else {
-                UpdateIsDirty();
             }
 
             DocumentContent = new RichTextViewModel(_fileInfo.CachedFilePath);
             DocumentContent.Content.PropertyChanged += (s,e) => UpdateHash();
             DocumentContent.ApiProvider.Save = SaveChanges;
+
+            UpdateIsDirty();
         }
 
         public RichTextViewModel DocumentContent { get; }
         public NotifyingProperty<bool> IsDirty { get; }
         
         public void Close() {
+            IsDirty.Value = false;
             ArgsAndSettings.CachedFiles.Remove(_fileInfo);
-            File.Delete(_fileInfo.CachedFilePath);
+            Directory.Delete(Path.GetDirectoryName(_fileInfo.CachedFilePath), true);
         }
 
         private void UpdateIsDirty() {
@@ -62,6 +75,17 @@ namespace NotepadSharp {
         private void SaveChanges() {
             if(IsDirty.Value) {
                 UpdateCachedFile();
+
+                if (_fileInfo.OriginalFilePath == null) {
+                    var dialog = new SaveFileDialog();
+                    dialog.DefaultExt = ".txt";
+                    dialog.Filter = "All Types|*.*";
+                    dialog.InitialDirectory = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
+                    if (!dialog.ShowDialog() ?? false) return;
+                    _fileInfo.OriginalFilePath = dialog.FileName;
+                    _updateLabelCallback(Path.GetFileName(dialog.FileName));
+                }
+                
                 File.Copy(_fileInfo.CachedFilePath, _fileInfo.OriginalFilePath, true);
                 IsDirty.Value = false;
             }
